created tica submodule
