+++ mlcvs/tica/__init__.py
@@ -1,0 +1,2 @@
+from .linear_tica import TICA_CV
+from .tica import TICA
